soc: hpe: gxp-xreg: fix format string type for GENMASK result
Cast the GENMASK(1, 0) result to unsigned int before passing to
%x format specifier. GENMASK returns unsigned long on ARM, causing
a -Wformat warning.

+++ drivers/soc/hpe/gxp-xreg.c
@@ -151,7 +151,8 @@ static ssize_t sideband_sel_show(struct device *dev,
 
 	regmap_read(drvdata->xreg_map, XREG_PSU_SIDEBAND, &value);
 
-	return sysfs_emit(buf, "0x%02x\n", value & XREG_PSU_SIDEBAND_MASK);
+	return sysfs_emit(buf, "0x%02x\n",
+			  (unsigned int)(value & XREG_PSU_SIDEBAND_MASK));
 }
 
 static ssize_t sideband_sel_store(struct device *dev,
